Resampling
This file resamples the features from different subsystems and generates two csv files; one for the beginning of the mission and one for the rest of the mission. In the beginning file, "Resample_Data.csv", the windows are constant time interval. In the second file, "Resample_Data_2.csv" each dark and light period creates a window.